root: not found file on delete images

+++ apps/root/src/file/file.service.ts
@@ -25,6 +25,10 @@ export class FileService {
       },
     });
 
+    if (!deletedFile) {
+      throw new GoneException('File not found');
+    }
+
     fs.unlink('./images/' + deletedFile.name, (err) => {
       if (err) {
         throw new GoneException('File not found');
